Exclude 'nightly' releases from Beta branch.
Excludes 'nightly' releases from version checks and updates the predicate for stable branch filtering to use a boolean instead of a string. Cleans up whitespace.

In the future, support for nightly might be added as its own branch but will require some changes to make sure we detect when a new nightly build is available.