Fix URL loading on tablets
Moved the URL loading logic from `makeUIView` to `updateUIView` within the `WebView` component.

This change prevents the `WKWebView` from reloading the same URL every time the view updates. A new property, `lastLoadedUrl`, was added to the `Coordinator` to track the currently displayed URL, ensuring that a new `URLRequest` is only loaded when the `url` property actually changes.

This lastLoadedUrl also makes sure to reload the webview page when the selected device changes on an iPad